Refactor routes implementing validation through front-end routing
