+++ src/routes/index.jsx
@@ -1,17 +1,36 @@
-import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
+
+import {jwtDecode} from "jwt-decode";
 
 import { AppRoutes as AppRoutesComponent } from './app.routes';
 import { AuthRoutes as AuthRoutesComponent } from './auth.routes';
-import { Register } from '../pages/Register';
 
 export function AppRoutes() {
-    return (
-        <BrowserRouter>
-            <Routes>
-                <Route path="/" element={<AuthRoutesComponent />} />
-                <Route path="/app" element={<AppRoutesComponent />} />
-                <Route path="/register" element={<Register />} />
-            </Routes>
-        </BrowserRouter>
-    );
+    // let token = localStorage.getItem('token');
+    let token = "dsahkjdhasjkdsa";  
+    const isTokenVerified = isTokenValid(token);
+
+    function isTokenValid(userToken) {
+        try {
+            jwtDecode(userToken)
+            return true;
+        } catch (err) {
+            console.log(err);
+            return false;
+        }
+    }
+
+    if (isTokenVerified) {
+        return (
+            <BrowserRouter>
+                <AppRoutesComponent />
+            </BrowserRouter>
+        );
+    } else {
+        return (
+            <BrowserRouter>
+                <AuthRoutesComponent />
+            </BrowserRouter>
+        );
+    };
 };
